+++ src/otx/utils/utils.py
@@ -12,6 +12,7 @@ from decimal import Decimal
 from typing import TYPE_CHECKING, Any, Callable
 
 if TYPE_CHECKING:
+    from pathlib import Path
     from types import FrameType
 
 
@@ -122,3 +123,31 @@ def get_decimal_point(num: float) -> int:
         return abs(exponent)
     error_msg = f"Can't get an exponent from {num}."
     raise ValueError(error_msg)
+
+
+def find_file_recursively(directory: Path, file_name: str) -> Path | None:
+    """Find the file from the direcotry recursively. If multiple files have a same name, return one of them.
+
+    Args:
+        directory (Path): directory where to find.
+        file_name (str): file name to find.
+
+    Returns:
+        Path | None: Found file. If it's failed to find a file, return None.
+    """
+    if found_file := list(directory.rglob(file_name)):
+        return found_file[0]
+    return None
+
+
+def remove_matched_files(directory: Path, pattern: str, file_to_leave: Path | None = None) -> None:
+    """Remove all files matched to pattern except file_to_leave.
+
+    Args:
+        directory (Path): direcetory to find files to remove.
+        pattern (str): pattern to match a file name.
+        file_not_to_remove (Path | None, optional): files to leave. Defaults to None.
+    """
+    for weight in directory.rglob(pattern):
+        if weight != file_to_leave:
+            weight.unlink()
